Succesfully retrieved all songs from Spotify
This is a temporary end of the spotify part for this project. Successfully retrieved names for all the songs in spotify, and next step is to add these songs to Youtube Music.

However, the spotify code isn't well structued. There are a few things to come back to once I finish migrating the songs to youtube.
1. How to automate the process of retrieving the auth code from the redirect uri.
2. Use OOP to make the code more managalbe.
3. Still having trouble with having to repeatedly authorize for use.Do more research in this for sure.